fix(build): resolve startup failures across all three backend services
- Remove duplicate version catalog registration from settings.gradle.kts
  (Gradle 8.11.1 auto-detects gradle/libs.versions.toml)
- Hardcode Spring Boot/Cloud BOM versions in root build.gradle.kts
  (libs extension inaccessible inside subprojects block)
- Add spring-boot-starter-web to libs:common for HttpStatus
- Add @EntityScan to all three Application classes to scan uz.kidscard.common
  (OutboxEvent lives outside each service's base package)
- Fix SMALLINT column type mismatches: expiryMonth/expiryYear in KidsCard,
  dayOfWeek/dayOfMonth in AllowanceSchedule — changed Int to Short with
  explicit columnDefinition = "SMALLINT" to match PostgreSQL DDL
- Add .toInt()/.toShort() conversions in CardService and CardDtos

+++ services/card-service/src/main/kotlin/uz/kidscard/card/api/dto/CardDtos.kt
@@ -22,8 +22,8 @@ data class IssueCardRequest(
 data class SetAllowanceRequest(
     @field:Min(1000) val amountUzs: Long,
     @field:NotNull val frequency: AllowanceFrequency,
-    val dayOfWeek: Int? = null,
-    val dayOfMonth: Int? = null,
+    val dayOfWeek: Short? = null,
+    val dayOfMonth: Short? = null,
 )
 
 // ── Responses ────────────────────────────────────────────────────────────────
@@ -48,8 +48,8 @@ data class AllowanceScheduleDto(
     val cardId: UUID,
     val amountUzs: Long,
     val frequency: AllowanceFrequency,
-    val dayOfWeek: Int?,
-    val dayOfMonth: Int?,
+    val dayOfWeek: Short?,
+    val dayOfMonth: Short?,
     val active: Boolean,
     val nextRunAt: Instant?,
 )
@@ -62,8 +62,8 @@ fun KidsCard.toDto() = KidsCardDto(
     familyId = familyId,
     cardType = cardType,
     maskedPan = maskedPan,
-    expiryMonth = expiryMonth,
-    expiryYear = expiryYear,
+    expiryMonth = expiryMonth.toInt(),
+    expiryYear = expiryYear.toInt(),
     network = network,
     status = status,
     balanceUzs = balanceUzs,
